requirements-gw.txt
aiohttp[speedups]
discord.py[speed] @ git+https://github.com/milenakos/discord.py-catbot
Pillow
Pilmoji
emoji==1.7
requests
discord-emoji
winuvloop
psutil
sentry-sdk
asyncpg
